I've added three html pages and have created a functioning nav bar that links the pages together. I've also added this nav bar to the footer of the window
The window is now scalable according to the size of teh screen. On a mobile, the logo stacks on top of the banner image. On a desk top it appears to the left of the banner image.

Added a lot more styling to CSS file

Created a contact page that allows the user to press a link and be redirected to their email. The to bar already has my email inputted.

I have also added a 'Things to do' page with an activity profile.